BS 54 | extracted common properties to generic and extend them for count and line report
