+++ src/main/java/org/bahmni/reports/model/TSIntegrationDiagnosisLineReportConfig.java
@@ -6,55 +6,9 @@ import java.util.List;
 
 
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class TSIntegrationDiagnosisLineReportConfig implements Config {
-    private String tsConceptSource;
-    private String conceptNameDisplayFormat;
-    private String terminologyParentCode;
-    private String terminologyColumnName;
-    private Boolean displayTerminologyCode;
+public class TSIntegrationDiagnosisLineReportConfig extends TSIntegrationDiagnosisReportConfig implements Config {
     private List<String> patientAttributes;
     private List<String> patientAddresses;
-
-    public String getTsConceptSource() {
-        return tsConceptSource;
-    }
-
-    public void setTsConceptSource(String tsConceptSource) {
-        this.tsConceptSource = tsConceptSource;
-    }
-
-    public String getConceptNameDisplayFormat() {
-        return conceptNameDisplayFormat;
-    }
-
-    public void setConceptNameDisplayFormat(String conceptNameDisplayFormat) {
-        this.conceptNameDisplayFormat = conceptNameDisplayFormat;
-    }
-
-    public String getTerminologyParentCode() {
-        return terminologyParentCode;
-    }
-
-    public void setTerminologyParentCode(String terminologyParentCode) {
-        this.terminologyParentCode = terminologyParentCode;
-    }
-
-    public Boolean isDisplayTerminologyCode() {
-        return displayTerminologyCode;
-    }
-
-    public void setDisplayTerminologyCode(Boolean displayTerminologyCode) {
-        this.displayTerminologyCode = displayTerminologyCode;
-    }
-
-    public String getTerminologyColumnName() {
-        return terminologyColumnName;
-    }
-
-    public void setTerminologyColumnName(String terminologyColumnName) {
-        this.terminologyColumnName = terminologyColumnName;
-    }
-
     public List<String> getPatientAttributes() {
         return patientAttributes;
     }
